Refactored navbar class so that every page with navbar can inherit from it
