+++ app/src/main/java/amazingme/activities/app/NavigationBarActivity.java
@@ -1,0 +1,141 @@
+package amazingme.activities.app;
+
+import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.NavigationView;
+import android.support.design.widget.Snackbar;
+import android.support.v4.view.GravityCompat;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.widget.TextView;
+
+import com.amazingme.activities.R;
+
+import amazingme.app.EnumeratedActivity;
+import amazingme.controller.ISessionLogoutHandler;
+import amazingme.model.AmazingMeAppCompatActivity;
+import amazingme.model.Parent;
+
+public abstract class NavigationBarActivity extends AmazingMeAppCompatActivity
+        implements NavigationView.OnNavigationItemSelectedListener, ISessionLogoutHandler {
+
+    public NavigationBarActivity(int layout) {
+        super(layout);
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        initializeNavigationBar();
+
+        // Register games when app is launched
+        goToIfNotSignedIn(EnumeratedActivity.LOGIN);
+    }
+
+    @Override
+    public void onBackPressed() {
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        if (drawer.isDrawerOpen(GravityCompat.START)) {
+            drawer.closeDrawer(GravityCompat.START);
+        } else {
+            super.onBackPressed();
+        }
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.main_menu, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.action_settings) {
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+
+    @SuppressWarnings("StatementWithEmptyBody")
+    @Override
+    public boolean onNavigationItemSelected(MenuItem item) {
+        // Handle navigation view item clicks here.
+        int id = item.getItemId();
+
+        if (id == R.id.nav_gallery) {
+
+        } else if (id == R.id.nav_slideshow) {
+
+        } else if (id == R.id.nav_manage) {
+
+        } else if (id == R.id.nav_share) {
+
+        } else if (id == R.id.nav_logout) {
+            getAppContext().sessionLogout(this);
+        } else if (id == R.id.nav_play) {
+            goTo(EnumeratedActivity.GAME_MENU);
+        }
+
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        drawer.closeDrawer(GravityCompat.START);
+        return true;
+    }
+
+    public void initializeNavigationBar() {
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
+                        .setAction("Action", null).show();
+            }
+        });
+
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        drawer.setDrawerListener(toggle);
+        toggle.syncState();
+
+        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
+        navigationView.setNavigationItemSelectedListener(this);
+
+        View header = navigationView.getHeaderView(0);
+        TextView nameText = (TextView) header.findViewById(R.id.nav_main_menu_name_text);
+        TextView emailText = (TextView) header.findViewById(R.id.nav_main_menu_email_text);
+
+        Parent parent = getUserContext().getParent();
+        if(parent != null) {
+            nameText.setText(parent.getDisplayName());
+            emailText.setText(parent.getEmail());
+        } else {
+            nameText.setText("");
+            emailText.setText("");
+        }
+    }
+
+    public void onSessionLogoutSuccess() {
+        goTo(EnumeratedActivity.LOGIN);
+    }
+
+    @Override
+    public void onSessionLogoutFailure(Exception e) {
+        // TODO: Add error handling when failing to doLogout
+    }
+
+}
