+++ app/src/main/res/layout/activity_game_menu.xml
@@ -1,36 +1,62 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android = "http://schemas.android.com/apk/res/android"
+<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools = "http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:layout_height = "match_parent" android:paddingLeft= "@dimen/activity_horizontal_margin"
-    android:paddingRight = "@dimen/activity_horizontal_margin"
-    android:paddingTop = "@dimen/activity_vertical_margin"
-    android:paddingBottom = "@dimen/activity_vertical_margin" tools:context = "amazingme.activities.app.LoginActivity">
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/drawer_layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:fitsSystemWindows="true"
+    tools:openDrawer="start">
 
-    <TextView
-        android:id="@+id/textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:text="Games"
-        android:textColor="@color/amazing_me_blue"
-        android:textSize="35dp"
-        android:textStyle="bold" />
-
-
-    <GridView
-        android:id="@+id/gridview"
+    <include
+        layout="@layout/app_bar_main_menu"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:numColumns="2"
-        android:verticalSpacing="30dp"
-        android:horizontalSpacing="20dp"
-        android:stretchMode="columnWidth"
-        android:gravity="center"
-        android:layout_below="@+id/textview"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        android:layout_marginTop="35dp" />
+        android:layout_height="match_parent" />
 
-</RelativeLayout>
+    <android.support.design.widget.NavigationView
+        android:id="@+id/nav_view"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:layout_gravity="start"
+        android:fitsSystemWindows="true"
+        app:headerLayout="@layout/nav_header_main_menu"
+        app:menu="@menu/activity_main_menu_drawer" />
+
+    <RelativeLayout xmlns:android = "http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools = "http://schemas.android.com/tools" android:layout_width="match_parent"
+        android:layout_height = "match_parent" android:paddingLeft= "@dimen/activity_horizontal_margin"
+        android:paddingRight = "@dimen/activity_horizontal_margin"
+        android:paddingTop = "@dimen/activity_vertical_margin"
+        android:paddingBottom = "@dimen/activity_vertical_margin" tools:context = "amazingme.activities.app.LoginActivity">
+
+
+        <TextView
+            android:id="@+id/textview"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerHorizontal="true"
+            android:text="Games"
+            android:textColor="@color/amazing_me_blue"
+            android:textSize="35dp"
+            android:textStyle="bold"
+            android:layout_marginTop="50dp" />
+
+
+        <GridView
+            android:id="@+id/gridview"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:numColumns="2"
+            android:verticalSpacing="30dp"
+            android:horizontalSpacing="20dp"
+            android:stretchMode="columnWidth"
+            android:gravity="center"
+            android:layout_below="@+id/textview"
+            android:layout_alignParentLeft="true"
+            android:layout_alignParentStart="true"
+            android:layout_marginTop="35dp" />
+
+    </RelativeLayout>
+</android.support.v4.widget.DrawerLayout>
+
